vamos ver
tomara

+++ src/modules/bjs_interpreter/interpreter.cpp
@@ -274,7 +274,7 @@ std::vector<Option> getScriptsOptionsList(String currentPath, bool saveStartupSc
                      String currentFolderName = currentPath.substring(currentPath.lastIndexOf('/') + 1);
                      String searchTitle = "[ " + currentFolderName + " ]";
 
-                     for (int i = 0; i < parentOptions.size(); i++) {
+                     for (size_t i = 0; i < parentOptions.size(); i++) {
                          if (parentOptions[i].label == searchTitle) {
                              restoreIndex = i;
                              break;
